backend/requirements.txt
# FastAPI y servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9
alembic==1.12.1

# Supabase
supabase==2.0.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# AI/ML
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.2
openai>=1.6.1
chromadb>=0.4.18
sentence-transformers>=2.2.2
scikit-learn>=1.4.0
pandas>=2.1.3
numpy>=1.26.2

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
httpx<0.25.0
